fix endless loop on certain input
When a modifier key in combination with a non alphanumeric key is
pressed, such as 'shift+tab', the program will hang. This is due to
the '_pos_read' variable not being increased, causing the while loop to
never break.

This fix adds a default case that returns a null input type and
increases the '_pos_read' variable accordingly.