improve resample function
This function is used to resample the bar count into a specific size
to fit the screen width. The pair contains the magnitude and the
corresponding frequency for each bar. This allows the ability to still
calculate the scientific musical notation represented by each bar
after a resample takes place.

The interpolation step adds linearly smoothed points inbetween each
existing point. The last point adds n number of copies of itself.

The decimation step reduces n number of points to 1 point by following
several steps. First, n number of points are sorted from greatest to
least by magnitude. Second, std::unique is used to remove all but the
first of every consecutive group of equivalent points. Third, in order,
each points frequency is compared to the previously choosen point,
selecting the first one that has a greater frequency or the last point
if none are greater. This is done to make sure a bar with a high
magnitude will be guaranteed to be visible when there is more bars than
screen width. It also decreases the chance of two consecutive bars that  represent the same musical note.